Internals

Internals

Kaleido.KaleidoLens — Type.

KaleidoLens <: Lens

Internal abstract type for Kaleido.jl.

source

Kaleido.SingletonLens — Type.

SingletonLens()

Inverse of FlatLens(1).

source

Kaleido._compose — Method.

_compose(lens1, lens2)

Like ∘ but fixes the associativity to match with the default one in Setfield.

source

Kaleido.prefer_singleton_callable — Method.

prefer_singleton_callable(f)

Convert f to an callable singleton object if possible. Useful if f is a Type.

Examples

julia> using Kaleido: prefer_singleton_callable

julia> sizeof((Int,))
8

julia> sizeof((prefer_singleton_callable(Int),))
0

julia> prefer_singleton_callable(Int)(1.0)
1

source